# System events performance improvement