--- a/server/src/main/java/io/spine/system/server/DefaultSystemWriteSide.java
+++ b/server/src/main/java/io/spine/system/server/DefaultSystemWriteSide.java
@@ -29,6 +29,7 @@ import io.spine.core.UserId;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.grpc.StreamObservers.noOpObserver;
 import static io.spine.system.server.SystemEventFactory.forMessage;
+import static java.util.concurrent.ForkJoinPool.commonPool;
 
 /**
  * The default implementation of {@link SystemWriteSide}.
@@ -55,8 +56,11 @@ final class DefaultSystemWriteSide implements SystemWriteSide {
         checkNotNull(systemEvent);
         checkNotNull(origin);
         Event event = event(systemEvent, origin);
-        system.eventBus()
-              .post(event, noOpObserver());
+        if (system.config().postEventsInParallel()) {
+            commonPool().execute(() -> postEvent(event));
+        } else {
+            postEvent(event);
+        }
         return event;
     }
 
@@ -64,5 +68,10 @@ final class DefaultSystemWriteSide implements SystemWriteSide {
         SystemEventFactory factory = forMessage(message, origin, system.isMultitenant());
         Event event = factory.createEvent(message, null);
         return event;
+    }
+
+    private void postEvent(Event event) {
+        system.eventBus()
+              .post(event, noOpObserver());
     }
 }