--- a/version.gradle
+++ b/version.gradle
@@ -25,7 +25,7 @@
  * as we want to manage the versions in a single source.
  */
 
-final def spineVersion = '1.5.4'
+final def spineVersion = '1.5.5'
 
 ext {
     // The version of the modules in this project.