… license-report.md
… server/src/test/java/io/spine/system/server/SystemSettingsTest.java