In this PR we allow system events to be posted in parallel. This improves performance in certain conditions.

The parallel publishing is enabled by default for production use and disabled for tests, since some tests rely on eventual consistency. It can be disabled for separate (or all) Bounded Contexts:
```java
BoundedContextBuilder builder = BoundedContext.singleTenant("example");
// ...
builder.systemFeatures().disableParallelPosting();
```

### Performance
For commands, we ran a test with 2000 commands posted from a single thread.
When system events are emitted synchronously, the test runs in 200-210 ms. After enabling parallelism, the runtime is shortened to 190-200 ms. 

For commands and imported events, the time of a synchronous run is close to 1030 ms, and for a paralleled run — 990 ms.

If the `CommandLog` system feature is enabled, the test runs in 970 ms in synch mode in paralleled mode — 970 ms.

Fixes #779.